% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.patterns.R
\name{make.patterns}
\alias{make.patterns}
\title{Construct collection of pattern specifications for tableplot}
\usage{
make.patterns(
  n = NULL,
  shape = 0,
  shape.col = "black",
  shape.lty = 1,
  cell.fill = "white",
  back.fill = "white",
  label = 0,
  label.size = 0.7,
  ref.col = "gray80",
  ref.grid = FALSE,
  scale.max = 1,
  as.data.frame = FALSE
)
}
\arguments{
\item{n}{Number of patterns}

\item{shape}{Shape(s) used to encode the numerical value of \code{cell}.
Any of \code{0="circle", 1="diamond", 2="square"}. Recycled to match the number of values
in the cell.}

\item{shape.col}{Outline color(s) for the shape(s)}

\item{shape.lty}{Outline line type(s) for the shape(s)}

\item{cell.fill}{inside color of |smallest| shape in a cell}

\item{back.fill}{background color of cell}

\item{label}{how many cell values will be labeled in the cell; max is 4}

\item{label.size}{size of cell label(s)}

\item{ref.col}{color of reference lines}

\item{ref.grid}{whether to draw ref lines in the cells or not}

\item{scale.max}{scale values to this maximum}

\item{as.data.frame}{whether to return a data.frame or a list.}
}
\value{
Returns either a data.frame of a list. If a data.frame, the pattern specifications appear as columns
}
\description{
Construct collection of pattern specifications for tableplot
}
\examples{
# None
}
