% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.colldiag.R
\name{tt.colldiag}
\alias{tt.colldiag}
\title{tinytable output method for colldiag objects}
\usage{
\method{tt}{colldiag}(x, dec.places = 3, fuzz = NULL, fuzzchar = ".", descending = FALSE, ...)
}
\arguments{
\item{x}{A \code{colldiag} object}

\item{dec.places}{Number of decimal places to use when printing}

\item{fuzz}{Variance decomposition proportions less than \emph{fuzz} are printed as \emph{fuzzchar}}

\item{fuzzchar}{Character for small variance decomposition proportion values}

\item{descending}{Logical; \code{TRUE} prints the values in descending order of condition indices}

\item{...}{arguments to be passed on to or from other methods (unused)

##@rdname colldiag}
}
\description{
This function uses tinytable to give a display of collinearity diagnostics with shaded backgrounds
}
\examples{
data(cars)
cars.mod <- lm (mpg ~ cylinder + engine + horse + weight + accel + year,
                data=cars)
(cd <- colldiag(cars.mod, center=TRUE))

tt(cd)


}
