% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableplot.colldiag.R
\name{tableplot.colldig}
\alias{tableplot.colldig}
\alias{tableplot.lm}
\alias{tableplot.glm}
\alias{tableplot.colldiag}
\title{Tableplot for Collinearity Diagnostics}
\usage{
\method{tableplot}{lm}(values, ...)

\method{tableplot}{glm}(values, ...)

\method{tableplot}{colldiag}(
  values,
  prop.col = c("white", "pink", "red"),
  cond.col = c("#A8F48D", "#DDAB3E", "red"),
  cond.max = 100,
  prop.breaks = c(0, 20, 50, 100),
  cond.breaks = c(0, 5, 10, 1000),
  show.rows = nvar:1,
  title = "",
  patterns,
  ...
)
}
\arguments{
\item{values}{A \code{"colldiag"}, \code{"lm"} or \code{"glm"} object}

\item{...}{other arguments, for consistency with generic}

\item{prop.col}{A vector of colors used for the variance proportions}

\item{cond.col}{A vector of colors used for the condition indices}

\item{cond.max}{Maximum value to scale the condition indices}

\item{prop.breaks}{Scale breaks for the variance proportions}

\item{cond.breaks}{Scale breaks for the condition indices}

\item{show.rows}{Rows to show in the display}

\item{title}{title used for the resulting graphic}

\item{patterns}{pattern matrix used for table plot.}
}
\description{
These methods produce a tableplot of collinearity diagnostics, showing the condition indices and variance
proportions for predictors in a linear or generalized linear regression model.
}
\examples{
# None yet

}
