% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.colldiag.R
\name{tt.colldiag}
\alias{tt.colldiag}
\title{\code{tinytable} Output Method for "colldiag" Objects}
\usage{
\method{tt}{colldiag}(
  x,
  digits = 2,
  fuzz = NULL,
  descending = FALSE,
  percent = FALSE,
  prop.col = c("white", "pink", "red"),
  cond.col = c("#A8F48D", "#DDAB3E", "red"),
  prop.breaks = c(0, 0.2, 0.5, 1),
  cond.breaks = c(0, 5, 10, 1000),
  font.scale = 1,
  ...
)
}
\arguments{
\item{x}{A \code{"colldiag"} object}

\item{digits}{Number of digits to use when printing; set to 0 when \code{percent = TRUE}}

\item{fuzz}{Variance decomposition proportions less than \emph{fuzz} are printed as \emph{fuzzchar}}

\item{descending}{Logical; \code{TRUE} prints the values in descending order of condition indices}

\item{percent}{Logical; if \code{TRUE}, the variance proportions are printed as percents, 0-100}

\item{prop.col}{A vector of colors used for the variance proportions. The default is \code{c("white", "pink", "red")}.}

\item{cond.col}{A vector of colors used for the condition indices, according to \code{cond.breaks}}

\item{prop.breaks}{Scale breaks for the variance proportions, a vector of length \emph{one more} than the number of \code{prop.col}, whose values are
between 0 and 1.}

\item{cond.breaks}{Scale breaks for the condition indices a vector of length \emph{one more} than the number of \code{cond.col}}

\item{font.scale}{Controls font size scaling for variance proportions. Either a single numeric value (no scaling) or a vector of length 2
specifying the minimum and maximum font sizes in \code{em} units. Default is \code{1} (no scaling). Use \code{c(1, 1.5)}
to scale font sizes from 1em to 1.5em based on variance proportion values, making larger proportions more visually prominent.}

\item{...}{arguments to be passed on to or from other methods (unused)}
}
\value{
a \code{"tinytable"} object
}
\description{
This function uses the \pkg{tinytable} package to give a display of collinearity diagnostics with
shaded backgrounds indicating the severity of collinearity in the dimensions of the data
and the proportions of variance related to each variable in these dimensions. It gives
a table version of the graphic shown by \code{\link{tableplot}}, but something that
can be rendered in HTML, PDF and other formats.
}
\details{
The \code{"tinytable"} object returned can be customized using other functions from
the \pkg{tinytable} package. For example, for HTML output, you can change the font family used in the output via \code{theme_html()}, as in:
\preformatted{
tt(cd) |>
  theme_html(css = "font-family: Arial, sans-serif;")
}
}
\examples{
library(VisCollin)
library(tinytable)
data(cars, package = "VisCollin")
cars.mod <- lm (mpg ~ cylinder + engine + horse + weight + accel + year,
                data = cars)
cd <- colldiag(cars.mod, center=TRUE)
# show all values, in same order as `cd`
tt(cd)

# show results in percent
tt(cd, percent = TRUE)

# try descending & fuzz
tt(cd, descending = TRUE, fuzz = 0.3)

# vary font size from 1em to 1.5em based on variance proportions
tt(cd, font.scale = c(1, 1.5))
}
\seealso{
\code{\link{colldiag}}, \code{\link{tableplot}},
\code{\link[tinytable]{tinytable}}
}
\author{
Michael Friendly
}
